Reordeded tests directories to use the same convention. Renamed tests files to use .test.js extension
